requirements.txt
ultralytics
opencv-python
requests>=2.33.1
python-dotenv
customtkinter
pillow>=12.2.0
psutil
nvidia-ml-py
openpyxl
lap
